Change how some values are derived during ingest.
Path patterns are different for the TS3 datasets so change
how we calculate the run number and a few other things. Remove
'raft' entry as there is no raft. Change 'ccd' to zero as there
is only one of them. We no longer seem to have jobIds so remove
this (obs_comCam never uses them from what I can remember anyway)

Put in a safety-check that the serial in the head matches that in
the path as I don't trust the info all that much. Also, a simple
assert here is no good, as it gets swallowed by a try block, whereas
putting in a raise with a message will show you why that file was
skipped.